Rename TestDomainHelpers.hpp to DomainTestHelpers.hpp
