+++ tests/Unit/Domain/Test_Domain.cpp
@@ -8,7 +8,7 @@
 #include "Domain/CoordinateMaps/CoordinateMap.hpp"
 #include "Domain/Domain.hpp"
 #include "Utilities/MakeVector.hpp"
-#include "tests/Unit/Domain/TestDomainHelpers.hpp"
+#include "tests/Unit/Domain/DomainTestHelpers.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 
 namespace {
